requirements.txt
pandas
numpy
scikit-learn
xgboost
torch
transformers
nltk
matplotlib
plotly
python-dotenv
requests
streamlit
Flask
Flask-Cors
websocket-client
angel-one-smartapi
python-telegram-bot
pushbullet.py
